+++ resources/js/components/bibliography/Citation.vue
@@ -18,8 +18,13 @@
 
 <script>
     import { computed } from 'vue';
+
+    import {
+        formatAuthors,
+        formatBibtexText,
+    } from '@/helpers/bibliography.js';
     import { truncate } from '@/helpers/filters.js';
-    import { formatAuthors } from '@/helpers/bibliography.js';
+
     import { showLiteratureInfo } from '@/helpers/modal.js';
     export default {
         props: {
@@ -30,7 +35,8 @@
         },
         setup(props) {
             const shortenedTitle = computed(() => {
-                return truncate(props.value.title, 60);
+                const formattedText = formatBibtexText(props.value.title);
+                return truncate(formattedText, 60);
             });
             return {
                 formatAuthors,