+++ resources/js/components/modals/bibliography/Details.vue
@@ -48,9 +48,9 @@
                     </dt>
                     <dd
                         class="col-md-9 font-monospace"
-                        :title="state.data.entry_type"
+                        :title="state.data.data.entry_type"
                     >
-                        {{ t(`main.bibliography.types.${state.data.entry_type}`) }}
+                        {{ t(`main.bibliography.types.${state.data.data.entry_type}`) }}
                     </dd>
                     <dt class="col-md-3 text-end">
                         {{ t('main.bibliography.column.author') }}